+++ src/components/zen-table-row/test/zen-table-row.spec.tsx
@@ -1,13 +1,18 @@
-let cells: NodeListOf<HTMLZenTableCellElement> = [];
+import { expect } from '@jest/globals';
+
+let cells: NodeListOf<HTMLZenTableCellElement>;
 
 import * as helpers from '../../helpers/helpers';
 helpers.getDefaultSlotContent = jest.fn(() => cells);
+
+import { mutationObserverMock, simulateMouse } from '../../helpers/jest';
+global.MutationObserver = mutationObserverMock();
 
 import { newSpecPage } from '@stencil/core/testing';
 import { ZenTableRow } from '../zen-table-row';
 import { ZenTable } from '../../zen-table/zen-table';
 import { ZenTableCell } from '../../zen-table-cell/zen-table-cell';
-import { simulateMouse } from '../../helpers/jest';
+import { cleanupTableStructure } from '../../zen-table/zen-table-helpers';
 
 describe('zen-table-row', () => {
   it('should render', async () => {
@@ -43,35 +48,40 @@ describe('zen-table-row', () => {
           </zen-table-row>
         </zen-table>`,
     });
+    const table = page.root as HTMLZenTableElement;
     const mainRow = page.root.querySelector('zen-table-row') as HTMLZenTableRowElement;
     cells = mainRow.querySelectorAll('zen-table-cell');
     mainRow.selectable = true;
     await page.waitForChanges();
+    await cleanupTableStructure(table);
 
-    expect(mainRow.visible).toBeTruthy();
+    expect(mainRow.$visible).toBe(true);
 
     const firstCell = page.root.querySelector('.first-cell') as HTMLZenTableRowElement;
 
     const expandRow = page.root.querySelector('[full-span]') as HTMLZenTableRowElement;
-    expect(expandRow.visible).toBeFalsy();
+    expect(expandRow.$visible).toBeFalsy();
 
     const expandableIcon = firstCell.shadowRoot.querySelector('.expand-icon');
     expect(expandableIcon).toBeTruthy();
 
     simulateMouse('click', expandableIcon);
-    expect(mainRow.expanded).toBeTruthy();
-    expect(expandRow.visible).toBeTruthy();
+    cleanupTableStructure(table);
+    expect(mainRow.expanded).toBe(true);
+    expect(expandRow.$visible).toBe(true);
   });
 });
 
 describe('zen-table-row tree functionality', () => {
   let page: SpecPage;
+  let table: HTMLZenTableElement;
   let parentRow: HTMLZenTableRowElement;
   let secondDepthParentRow: HTMLZenTableRowElement;
   let firstCell: HTMLZenTableCellElement;
   let secondDepthCell: HTMLZenTableCellElement;
 
   beforeEach(async () => {
+    global.MutationObserver.mockClear();
     page = await newSpecPage({
       components: [ZenTable, ZenTableRow, ZenTableCell],
       html: /*html*/ `
@@ -96,25 +106,27 @@ describe('zen-table-row tree functionality', () => {
           </zen-table-row>
         </zen-table>`,
     });
-    parentRow = page.root.querySelector('zen-table-row') as HTMLZenTableRowElement;
-    secondDepthParentRow = page.root.querySelector('[second-level-parent]') as HTMLZenTableRowElement;
+    table = page.root;
+    parentRow = table.querySelector('zen-table-row') as HTMLZenTableRowElement;
+    secondDepthParentRow = table.querySelector('[second-level-parent]') as HTMLZenTableRowElement;
 
     // mock slots:
     cells = parentRow.querySelectorAll('zen-table-cell');
     parentRow.selectable = true;
-    parentRow.expandable = true;
+    parentRow.$expandable = true;
     parentRow.expanded = true;
 
-    firstCell = page.root.querySelector('.first-cell') as HTMLZenTableCellElement;
-    secondDepthCell = page.root.querySelector('.second-depth-cell') as HTMLZenTableCellElement;
+    firstCell = table.querySelector('.first-cell') as HTMLZenTableCellElement;
+    secondDepthCell = table.querySelector('.second-depth-cell') as HTMLZenTableCellElement;
     secondDepthCell.$selectable = true;
 
     await page.waitForChanges();
   });
 
   it('should set all descendants to not visible on expand false', async () => {
-    expect(parentRow.visible).toBeTruthy();
-    expect(secondDepthParentRow.visible).toBeTruthy();
+    cleanupTableStructure(table);
+    expect(parentRow.$visible).toBe(true);
+    expect(secondDepthParentRow.$visible).toBe(true);
 
     const expandableIcon = firstCell.shadowRoot.querySelector('.expand-icon');
 
@@ -125,7 +137,7 @@ describe('zen-table-row tree functionality', () => {
     descendents.concat(page.root.querySelectorAll('zen-table-row[depth="2"]'));
 
     for (const row of descendents.values()) {
-      expect((row as HTMLZenTableRowElement).visible).not.toBeTruthy();
+      expect((row as HTMLZenTableRowElement).$visible).not.toBe(true);
     }
   });
 
@@ -137,13 +149,13 @@ describe('zen-table-row tree functionality', () => {
     descendents.concat(page.root.querySelectorAll('zen-table-row[depth="2"]'));
 
     for (const row of descendents.values()) {
-      expect((row as HTMLZenTableRowElement).selected).toBeTruthy();
+      expect((row as HTMLZenTableRowElement).selected).toBe(true);
     }
 
     parentRow.selected = false;
 
     for (const row of descendents.values()) {
-      expect((row as HTMLZenTableRowElement).selected).not.toBeTruthy();
+      expect((row as HTMLZenTableRowElement).selected).not.toBe(true);
     }
   });
 
@@ -157,10 +169,10 @@ describe('zen-table-row tree functionality', () => {
     secondDepthParentRow.selected = true;
 
     for (const row of firstDepthChildren.values()) {
-      expect((row as HTMLZenTableRowElement).selected).not.toBeTruthy();
+      expect((row as HTMLZenTableRowElement).selected).not.toBe(true);
     }
     for (const row of secondDepthChildren.values()) {
-      expect((row as HTMLZenTableRowElement).selected).toBeTruthy();
+      expect((row as HTMLZenTableRowElement).selected).toBe(true);
     }
   });
 
@@ -170,8 +182,9 @@ describe('zen-table-row tree functionality', () => {
     const event = new Event('change', { bubbles: true, composed: true });
     checkbox.dispatchEvent(event);
     await page.waitForChanges();
+    cleanupTableStructure(table);
 
-    expect(parentRow.$indeterminate).toBeTruthy();
-    expect(secondDepthParentRow.$indeterminate).toBeTruthy();
+    expect(parentRow.$indeterminate).toBe(true);
+    expect(secondDepthParentRow.$indeterminate).toBe(true);
   });
 });