+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+## [3.3.1](https://github.com/reciprocity/zen-ui/compare/3.3.0...3.3.1) (2021-04-13)
+
+
+### Bug Fixes
+
+* first row should have top blue border when selected ([a250dfc](https://github.com/reciprocity/zen-ui/commit/a250dfc12d31d090db45f4968294e7dd7add55da))
+
+# [3.3.0](https://github.com/reciprocity/zen-ui/compare/3.2.5...3.3.0) (2021-04-13)
+
+
+### Features
+
+* **zen-card:** add additional shadow variant ([2102c61](https://github.com/reciprocity/zen-ui/commit/2102c6140513dcbadb2de10e7774aafe39543198))
+
+## [3.2.5](https://github.com/reciprocity/zen-ui/compare/3.2.4...3.2.5) (2021-04-12)
+
+
+### Bug Fixes
+
+* **zen-progress-tracker:** labels now have white-space nowrap style ([8fe737c](https://github.com/reciprocity/zen-ui/commit/8fe737c84ce252e4c0c61553be1650133f65e5c0))
+
+## [3.2.4](https://github.com/reciprocity/zen-ui/compare/3.2.3...3.2.4) (2021-04-12)
+
+
+### Bug Fixes
+
+* **zen-tabs:** fix that tabs stretch full width ([995f32e](https://github.com/reciprocity/zen-ui/commit/995f32e99506808f5eb004aabcd9291ce201ab78))
+
+## [3.2.3](https://github.com/reciprocity/zen-ui/compare/3.2.2...3.2.3) (2021-04-12)
+
+
+### Bug Fixes
+
+* non-selectable and non-expandable child row should still have left padding ([898120d](https://github.com/reciprocity/zen-ui/commit/898120d353c80fb589833c8b7d3097d56390159e))
+* **zen-table:** expandable and visible props are always correctly calculated ([6a7706a](https://github.com/reciprocity/zen-ui/commit/6a7706a8242afbfe24779c88a3928ef4458ceb1e))
+* **zen-table:** selected and indeterminate props are always correctly calculated ([feb73b9](https://github.com/reciprocity/zen-ui/commit/feb73b915f1e3980be20b919ded4974826c8cb52))
+
+## [3.2.2](https://github.com/reciprocity/zen-ui/compare/3.2.1...3.2.2) (2021-04-12)
+
+
+### Bug Fixes
+
+* **zen-space:** vertically distributed items should not have right margin ([db94404](https://github.com/reciprocity/zen-ui/commit/db94404714d890dc3679a8085333eb1ed9a4ee53))
+
 ## [3.2.1](https://github.com/reciprocity/zen-ui/compare/3.2.0...3.2.1) (2021-04-09)